Transform date/time format string from

As per the csvw specification for date and time formats we accept format strings using the date field symbols defined in unicode TR35. These are converted to POSIX 1003.1 date format strings for use in base::strptime() or readr::parse_date()/readr::parse_datetime().

transform_datetime_format(format_string)

Arguments

format_string
a UAX35 date format string

Value

a POSIX date format string

Examples

if (FALSE) {
fmt <- transform_datetime_format("dd.MM.yyyy")
strptime("01.01.2001", format=fmt)
}